Minimum set of changes required to get AK to compile on Aarch64
