+++ AK/Weakable.h
@@ -12,7 +12,7 @@
 #include "RefPtr.h"
 #include "StdLibExtras.h"
 #ifdef KERNEL
-#    include <Kernel/Arch/x86/Processor.h>
+#    include <Kernel/Arch/Processor.h>
 #    include <Kernel/Arch/x86/ScopedCritical.h>
 #endif
