+++ Kernel/Arch/x86/common/ASM_wrapper.cpp
@@ -7,7 +7,7 @@
 #include <AK/Types.h>
 
 #include <Kernel/Arch/x86/ASM_wrapper.h>
-#include <Kernel/Arch/x86/Processor.h>
+#include <Kernel/Arch/Processor.h>
 #include <Kernel/Sections.h>
 
 namespace Kernel {
