+++ Kernel/Arch/x86/common/Processor.cpp
@@ -10,7 +10,6 @@
 #include <AK/Types.h>
 
 #include <Kernel/Interrupts/APIC.h>
-#include <Kernel/Memory/ScopedAddressSpaceSwitcher.h>
 #include <Kernel/Process.h>
 #include <Kernel/Sections.h>
 #include <Kernel/StdLib.h>
@@ -19,11 +18,15 @@
 #include <Kernel/Arch/x86/CPUID.h>
 #include <Kernel/Arch/x86/Interrupts.h>
 #include <Kernel/Arch/x86/MSR.h>
-#include <Kernel/Arch/x86/Processor.h>
+#include <Kernel/Arch/Processor.h>
 #include <Kernel/Arch/x86/ProcessorInfo.h>
 #include <Kernel/Arch/x86/SafeMem.h>
 #include <Kernel/Arch/x86/ScopedCritical.h>
 #include <Kernel/Arch/x86/TrapFrame.h>
+#include <Kernel/Arch/x86/InterruptDisabler.h>
+
+#include <Kernel/Memory/ScopedAddressSpaceSwitcher.h>
+#include <Kernel/Memory/PageDirectory.h>
 
 namespace Kernel {
