+++ Kernel/Arch/x86/common/TrapFrame.cpp
@@ -5,6 +5,8 @@
  */
 
 #include <Kernel/Arch/x86/TrapFrame.h>
+#include <Kernel/Arch/x86/InterruptDisabler.h>
+#include <Kernel/Arch/Processor.h>
 
 namespace Kernel {
