+++ Kernel/Library/ThreadSafeNonnullRefPtr.h
@@ -12,7 +12,7 @@
 #include <AK/Traits.h>
 #include <AK/Types.h>
 #ifdef KERNEL
-#    include <Kernel/Arch/x86/Processor.h>
+#    include <Kernel/Arch/Processor.h>
 #    include <Kernel/Arch/x86/ScopedCritical.h>
 #endif
