+++ Kernel/Library/ThreadSafeRefPtr.h
@@ -15,7 +15,7 @@
 #include <AK/Types.h>
 #ifdef KERNEL
 #    include <Kernel/API/KResult.h>
-#    include <Kernel/Arch/x86/Processor.h>
+#    include <Kernel/Arch/Processor.h>
 #    include <Kernel/Arch/x86/ScopedCritical.h>
 #endif
