+++ Kernel/Panic.cpp
@@ -6,7 +6,7 @@
 
 #include <AK/Format.h>
 #include <Kernel/Arch/x86/IO.h>
-#include <Kernel/Arch/x86/Processor.h>
+#include <Kernel/Arch/Processor.h>
 #include <Kernel/CommandLine.h>
 #include <Kernel/KSyms.h>
 #include <Kernel/Panic.h>
