+++ Kernel/Random.cpp
@@ -6,7 +6,7 @@
  */
 
 #include <AK/Singleton.h>
-#include <Kernel/Arch/x86/Processor.h>
+#include <Kernel/Arch/Processor.h>
 #include <Kernel/Devices/RandomDevice.h>
 #include <Kernel/Random.h>
 #include <Kernel/Sections.h>
