+++ Kernel/UBSanitizer.cpp
@@ -6,7 +6,7 @@
 
 #include <AK/Format.h>
 #include <AK/UBSanitizer.h>
-#include <Kernel/Arch/x86/Processor.h>
+#include <Kernel/Arch/Processor.h>
 #include <Kernel/KSyms.h>
 
 using namespace Kernel;
